backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
gitpython==3.1.43
apscheduler==3.10.4
pydantic==2.7.1
lxml==5.2.2
curl-cffi==0.7.1
